requirements.txt
click==3.3
requests==2.7.0
PyYAML==3.11
